Fix false positive in NoUnused.Variables
Fixes #53

The issue was that the parameters from let/lambda functions and patterns from
case declarations were not taken into account. When we exit a scope, we
look at the variables introduced in the scope and the used variables
inside the scope. Used variables that are found in the
parameters/patterns don't get sent back to the parent scopes, while
those that are not found are propagated that way.

In this case, the root problem was that since the parameters were not
registered for these cases, we were not preventing used variables from
being propagated back to the top as we should have.

This change does 2 things:
- Use Rule.withLetDeclaration*Visitor and Rule.withCaseBranch*Visitor
  newly introduced in jfmengels/elm-review v2.6.0 (which gets rid of
  some complexity about when we enter a scope or not)
- Properly suppress the used variables (by ignoring some names, as we
  don't want to report those)

Thanks to @stefan-wullems for reporting the issue and for the helpful SSCCEs!